refactor: 이미지 폴더 경로 변경
