+++ src/app/review/complete/page.tsx
@@ -47,7 +47,7 @@ function ReviewComplete() {
           className="flex flex-col w-72 h-[500px] bg-gray-100 drop-shadow-xl rounded-b-md dark:text-black">
           <div className="flex justify-center mb-7">
             <Image
-              src="/assets/logo.webp"
+              src="~/images/logo.webp"
               width={100}
               height={100}
               alt="app logo"
@@ -96,7 +96,7 @@ function ReviewComplete() {
           <div className="flex flex-col items-center gap-2 mt-auto mb-3 pt-2 border-t-2 text-sm">
             <p>HandsUp에서 좋은 거래하세요 :)</p>
             <Image
-              src="/assets/barcode.webp"
+              src="/assets/images/barcode.webp"
               width={130}
               height={18}
               alt="barcode image"
